+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Create Wiki documentation ([Issue #31](https://github.com/crazy-max/neard/issues/31))
 * Move Console icons to core resources
 * Upgrade Console tool to pack 2 release 3
+* Retrieve latest version from wiki page `latestVersion.md` instead of `CHANGELOG.md`
 
 ### Upgrade from previous release